Remove test util and relocate contents to core test
